Oppdaterer init-sjekker for meldekortbehandlinger (WIP)
